+++ src/app/shared/EventInfo.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import cx from "classnames"
   import {Tags} from "paravel"
   import {Naddr} from "src/util/nostr"
   import {secondsToDate, formatTimestamp, formatTimestampAsDate} from "src/util/misc"
@@ -6,14 +7,20 @@
   import GroupLink from "src/app/shared/GroupLink.svelte"
   import PersonLink from "src/app/shared/PersonLink.svelte"
   import EventActions from "src/app/shared/EventActions.svelte"
+  import NoteContentKind1 from "src/app/shared/NoteContentKind1.svelte"
   import {router} from "src/app/router"
+  import {isDeleted, getSetting} from "src/engine"
 
   export let event
+  export let showDate = false
+  export let showActions = false
 
   const timeFmt = new Intl.DateTimeFormat("en-US", {timeStyle: "short"})
+  const datetimeFmt = new Intl.DateTimeFormat("en-US", {dateStyle: "short", timeStyle: "short"})
   const groupAddrs = Tags.from(event).circles().all()
   const address = Naddr.fromEvent(event).asTagValue()
-  const {name, start, end, location} = Tags.from(event).getDict()
+  const detailPath = router.at("events").of(address).toString()
+  const {title, start, end, location} = Tags.from(event).getDict()
   const startDate = secondsToDate(start)
   const endDate = secondsToDate(end)
   const startDateDisplay = formatTimestampAsDate(start)
@@ -22,11 +29,15 @@
 </script>
 
 <div class="flex flex-grow flex-col gap-2">
-  <div class="flex justify-between">
-    <Anchor class="text-2xl" href={router.at("events").of(address).toString()}>{name}</Anchor>
-    <EventActions {event} />
+  <div class="flex items-center justify-between">
+    <Anchor class={cx("text-2xl", {"line-through": $isDeleted(event)})} href={detailPath}>
+      {title}
+    </Anchor>
+    {#if showActions}
+      <EventActions {event} />
+    {/if}
   </div>
-  <div class="flex gap-2 text-sm text-lighter">
+  <div class="flex items-center gap-2 text-sm text-lighter">
     {#if event.wrap}
       <span>Private</span>
       <span>•</span>
@@ -50,11 +61,18 @@
   <div class="h-px bg-mid" />
   <div class="flex items-center gap-2 text-sm text-lighter">
     <i class="fa fa-clock" />
-    Starts at {timeFmt.format(startDate)} — {isSingleDay
-      ? timeFmt.format(endDate)
-      : formatTimestamp(end)}
+    {#if showDate}
+      Starts on {datetimeFmt.format(startDate)} — {isSingleDay
+        ? timeFmt.format(endDate)
+        : formatTimestamp(end)}
+    {:else}
+      Starts at {timeFmt.format(startDate)} — {isSingleDay
+        ? timeFmt.format(endDate)
+        : formatTimestamp(end)}
+    {/if}
     <span class="w-2" />
     <i class="fa fa-location-dot" />
     {location || "No location"}
   </div>
+  <NoteContentKind1 showEntire showMedia={getSetting("show_media")} note={event} />
 </div>