+++ src/app/shared/NoteContentKind31923.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
   import EventInfo from "src/app/shared/EventInfo.svelte"
-  import NoteContentKind1 from "src/app/shared/NoteContentKind1.svelte"
 
-  export let note, showMedia
+  export let note
 </script>
 
-<div class="flex flex-col gap-4 overflow-hidden text-ellipsis">
-  <EventInfo event={note} />
-  <NoteContentKind1 {note} {showMedia} />
-</div>
+<EventInfo showDate event={note} />